requirements.txt
streamlit==0.61.0
opencv-contrib-python
numpy==1.16.5
pillow==7.2.0
scikit-image
scipy